+++ CHANGELOG.md
@@ -3669,3 +3669,4 @@
 - [2026-08-17 09:58] test(state): add unit tests for state business logic
 - [2026-08-17 11:32] fix(security): correct layout constraints and overflow in security
 - [2026-08-17 12:41] test(test): add unit tests for test business logic
+- [2026-08-17 15:41] test(navigation): implement widget integration tests for navigation